requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pypdf
chromadb
sentence-transformers
openai>=1.40.0
